1114 유닛, 카메라 제어 스크립트 중간 저장

+++ Assets/Scripts/Common/Interface/IDamage.cs
@@ -1,5 +1,5 @@
 
 public interface IDamage 
 {
-    public void TakeDamage(int dmg);
+    public void TakeDamage(float dmg);
 }
